Adjust coding structure of tool-currencies
